fix(docs): move the camera auto-rotate helper into src/renderer
CI runs vitest without docs/node_modules, so a test importing a file under
docs/src/ fails when Vite resolves docs/tsconfig.json (its `extends` points
at @docusaurus/tsconfig). The helper now lives at
src/renderer/cameraAutoRotate.ts next to the controls it complements, is
imported by the docs components the same way MoleculeRenderer is, and its
test moves to tests/ts/renderer/ (plus a case for the default
requestAnimationFrame path).
